requirements.txt
numpy
pandas
GitPython
hydra-core
tqdm
tensorboard
scipy
Cython==0.29.35
torch==2.0.1
torch_geometric==2.4.0
torch-scatter==2.1.1